Updated Neon SDK and improved DefaultValueAttribute retrieval
Updated the Neon SDK version in the Build.cs file. Enhanced the method of retrieving the `DefaultValueAttribute` in the RoslynExtensions.cs file to search through all custom attributes. Added the `System.Runtime.Serialization` namespace to the RoslynExtensions.cs and Test_RoslynType.cs files. Added `DefaultValue` attributes to new properties in the `Bar` class and to the `Baz` property in the `Foo` class in the Test_RoslynType.cs file. Also added a new enum `MyEnum` with `EnumMember` attributes to the Test_RoslynType.cs file. Updated the expected output of the JSON serialization in the Test_RoslynType.cs file. Removed the `Publish Neon.ModelGenerator` line from the neonsdk-nuget-github.ps1 file.