Add ReadyFunction to ProxyConfig and update service readiness
Added a new using directive for System.Threading.Tasks in ProxyConfig.cs.
Introduced a ReadyFunction property in ProxyConfig to hold a function
that will be called when the service is ready. Modified ServiceDiscovery
to check and await ReadyFunction if it is not null. Updated Program.cs
in TestBlazor to set the ReadyFunction property with a delay of 1000 ms.